ZeN-4452 Working base api and openapi3-doc
These are positioned in the repo in an anticipated module structure, in
which all projects are with a modules directory. A subdirectory,
modules/gentemplates, will hold all the reprezen-provided gentemplates,
while the api module will appear separately.

We'll use the maven assembly plugin to create an uber-jar of all the
gentemplates to be deployed alongside the api jar.