ZEN-4452 Fix problems with shader config
We need to use a resource transformer for the META-INF/services files
that are scanned by the service loader at runtime to locate available
GenTemplates. Each GenTemplate project will have one, and we need them
all. Fortunately, the resource appender implementation that comes with
the shader jar is just fine for this.

Other resources need to be split out into folders named for the
project's main packages, now that we can no longer rely on the isolation
provided by bundle classloaders. So this has been done for the
OpenApi3Doc gentemplate.