ZEN-4453 FIx edge case with static resources
Code was failing to locate resources in plugin bundles, if the paths had
spaces in them. This can happen with built-in generators in workspaces
that themselves have spaces in their path. The bundle class loaders
apparently use bare spaces, not encoded spaces (%20) when reporting the
code location, while other class loaders encode the spaces. So when the
toURI method fails in the old code, that's caught, and new code just
uses the URL path to create a File.