ZEN-4454 Drop the whole uber-jar thingy
Turns out publishing a bunch of jar files simultaneously to maven
central is a piece of cake, assuming they're all part of a single
aggregate. So the primary motivation for the uber-jar is gone. Plus, the
shader was locking down versions for dependencies that we wanted to be
ranges, and it's not clear that it would be easy to solve
that. So... now everything's a fair bit simpler.

The `standard-gentemplates` module now gets pom packaging and is simply
a collection of dependencies for all available gentemplates as
individual artifacts. (Just one for now.)