ZEN-4458 Create tool to test gentemplates
With the build cycle so excrutiatingly long for API Studio (build
genflow, build target platform bundle, reload target platform, rebuild
API Studio, launch, test, go back and fix whatever was wrong and do it
all again...), I took some time to create a simple manual test rig for
genflow gentemplates, in the form of an executable jar built in
genflow-tools project. It's skipped for installation/deploymnet.

Give it a glob pattern as a command line arg. It lists the GenTemplates
whose ids match the glob, and asks you to select one of them by typing
in its number. It then invokes the gentemplate with a sample model of
the appropriate type.

A few minor repairs were done on genflow-api. Most significant is that
GenTemplateRegistry#getGenTemplate(String) now always creates a new
instance of the requested GenTemplate, since GenTemplate instances are
not reusable.