ZEN-4458 Dramatically improved GenTool
Now you can specify parameters and primary source. Plus, the existing
logic has been simplified.

For params you can choose one of the declareada parameter names or type
your own name. For values you can choose whether you want to enter a
string, an "empty string", a number, a boolean, or a JSON value. Empty
string is required because if you just type RETURN when entering a
string, it's interpreted as canceling the param input.

For primary source, a default is chosen for you based on the model
type. You can also put in your full path. If you do the latter, the
directory containing the file you choose becomes the "default" for next
time, and you're presented with all files with a conforming extensions
when you next ask to choose a file. (You can also just type a directory
path in this case, and that will be scanned immediately and you'll be
asked to select from that list).