ZEN-4458 Added RAML gentemplate (currently suppressed)
The RAML gentemplate was suppressed in API Studio, and it's now
suppressed in GenFlow. It also appears to be broken - throws NPE
somewehre in JsonSchema code.

That code was part of the JsonSchema GenTemplate, and it's used heavily
by the various Swagger gentemplates. Since it's used by multiple
gentemplates, I've moved it into genflow-common, trying to keep
target-specific variations elswhere. (But perhaps not succeeding when it
comes to XChange generation, since some XChange-related code seems to be
in the base classes)

The gentemplate also appears to be broken in current release. So it's
possible I screwed things up when copying over the JsonSchema stuff. Or
it could be that we broke it somewhere along the way by making changes
to the JS stuff without breaking the other gentemplates that rely on
it. I guess I'll find out when I migrate those other GenTemplates...