ZEN-4493 Change to GenTemplateRegistry
These changes are to allow GenTemplateRegistry to act as the source of
GenTemplate information for EclipseGenTemplateRegistry in API Studio.

As part of this, the really awful GenTemplateInfo class in API Studio is
replaced by a much more reasonable GenTemploteInfo claass in API
Studio. The main point of this was to set things up for eventually
making the interaction between EGTR and GTR an arms-length interaction,
transmitting serialized gentemplate descriptors, without actual
instances. This would hopefully set the stage for removing most of
GenFlow from the API Studio bundles plugin, signiifcantly reducing the
product size.

At this point, instantiation of a new GenTemplate instance is
problematic because some gentemplates (OAG and SCG) don't have no-arg
constructors. So we'll probably need some sort of factory approach
instead, so we can have a universal way of consing up an instance from
an Info object.

The changes to GenTemplateInfo had fairly widespread impact. Almost
nothing of significance, mostly just renaming of packages, methods,
classes, etc. I did take the opportunity of going through some of the UI
bits and largely redoing the unbelievably confusing names that have
always made those classes very difficult to work with. Hopefully this is
an improvement that will pay dividends in the future...