ZEN-4513 Use caller-supplied classloader in GTR
GTR instances can now be created with a caller-supplied classloader,
which will be used for scans. The static instance is now called the
"default" instance, and methods that use it also incorporate "default"
in their names.

This was important for EclipseGenTemplateRegistry, so gentemplates
present in the workspace would be included.