ZEN-4481 Fix content assembly logic in Normalizer
The `ContentLoader` class was recently augmented to recursively assemble
content from non-conforming references. Prevoiusly, the direclty
referenced content tree was loaded, but references within that tree were
not themselves loaded.

Of course, this implementation included a reference cycle test, which
used a set of references. WHen a reference was followed, it was added to
the set, as long as it was not already in the set (if it was then it's a
cyclic reference). And of course, after that, the reference was removed
form the set. Except it wasn't. That was left out. So any referenced
content that included hte same reference in multiple places would
properly resolve the first instance encountered, and it would claim the
others were cyclic.

This commit adds the ref removal when popping out of a reference.