ZEN-4495 Fix OpenApiGenerator
v3 models were not being normalized on the way into OpenAPI
Generator. This was the case when OAG was initially introduced into API
Studio, since normalizer at that time did not handle v3 models. Once
that support was introduced, the code in the OAG adapter was not updated
to take advantage of it. This commit does that.