Don't let a single bad gentemplate kill the scan
Scans for individual gentemplates and for gentemplate groups were both
wrong, in different ways.

For individual scans, we were catching ServiceConfigError and skipping
the problematic gentemplate with a warning. That was missing exceptions
that could occur during creation of a GenTemplateInfo object from the
discovered GenTEemplate.

In the group scan we were catching Exception, but ServiceConfigError is
not an Exception, so a problematic group could likewise kill the whole
scan.

Now both scans catch Throwable, so no problematic single discovery
should be capable of killing everything.