Retain implicitly required security schemes
This is an initial and limited implementation of something that was
missing from the design of the normalizer. Namely, may be required for
completeness in a model without having any reference nodes referring to
them. An example is showing up in the beamup example now that we're
using a KZOP version with more complete validations. The model contains
a security scheme that must be retained because its name appears in a
security requirement in the model.

This implementation fixes the live doc view for the beamup example, but
is incomplete in two ways:
1. There are other implicit retention requirements that are not
   implemented, e.g for callbacks and links
2. Top-level security requirements (i.e. not cotnained in an operation
   object) are not processed.