ZEN-4598 Fix resolution of jackson's yaml module
OAG generators use features of the `jackson-dataformat-yaml` module taht
do not appear in the version of that module resolved by maven in model
projects. This change causes this dependency to be managed from the
top-level pom file and also includes an explicit (managed) dpeendency in
standard-gentempaltes module. Net result is that v2.9.5 is used.