Add dependencies for commons-text
It's a mystery why prior versions were able to compile. A class in
genflow-common imported and used a class from apache's commons-text
library, but it was not required in the pom.xml and was not transitively
included. The swagger-doc gentemplate project had the same problem.

Adding the dependnecy to genflow-common fixed both projects (since
swagger-doc picked it upt transitively), but I've added it to both for
hygiene.