ZEN-4408 Improve ModuleInfo maintenance approach
Previously, a single pair of files - modulesInfo.csv and
moduleParams.csv - contained metadata extracted from SCG modules, as
well as any manual overrides (suppressions, name overrides, etc.) added
by us. This was problematic because the same files served as both input
and output in the discovery process, which could cause considerable
confusion, especially when manual updates were spread across multiple
commits with intervening builds.

The approach now is to use the same files but with the SCG version
string incorporated into the file name, like
modulesInfo_2.3.1.csv. These files are created by discovery only once,
the first time a given version is configured into the GenFlow
dependencies. At runtime, it's possible that no module data will be
available for the SCG version being used, and in that case, the model
info for the highest version not exceeding that version is used instead,
with "default" info collected for new codegen modules.

Now, discovery is a no-op when module info for the current version
already exists. Otherwise it finds "base" info file using the same
approach as above (highest version not exceeding current) and then
creates a file for the new version, carrying some data forward from the
base info.

New flags in the csv file include:
* New - set when an SCG module appears in the current version but did
  not appear in the base file
* Vetted - can only be set manually, indicates that this SCG module has
  been checked for proper naming, and an override has been provided if
  needed.

In addition, the "discovered" flag has been renamed to "builtin" - still
not a great name, but at runtime it distinguishes built-in modules from
custom modules provided by the user.

At runtime, only vetted built-in modules and custom modules (necessarily
unvetted) are made available as GenTemplates.

The biggest challenge in this change was due to the need to search for
existing module info files, rather than targeting a single file. This
required code to search for files for the three ways base locations can
be specified: file: URL, jar: URL, and budleresource: URL.