Update pom files so `mvn clean install` works
Pom files for test projects that access Rapid models contained in the
genflow-test project are updated. They now have a
`maven-dependencies-plugin` build excecution that executes in
`process-test-resources` phase. The execution explodes all `.rapid`
files from the `genflow-test` jar file into the current project's
`target/test-classes` directory. These will be used when tests are run,
because that the classpath entry for that directory will precede the
`genflow-tests` jar file.

This is needed because the `ZenModelParser` is not able to deal with the
`jar:` URLs that are used to address resources found in jar files.