[ZEN-4725] GenFlow should follow Maven lifecycle, should not require install before test
This commit allows tests to be run during maven install phase. To do that the method getRestFileURI in GeneratorTestFixture that returns the URI of a test file has been modified to allow the use of archive based URIs (starting with jar:) as well as file based URIs (starting with file://). This allows to run tests within Eclipse as well as running them during maven install phase (where archive URI will be used).

This also requires a change in RAPID-ML project that allows resolution of imports when those are present in archive files. This has been done in a seprate PR that basically discard use of java.net.URI for EMF URI which allow resolving of archive URIs whcih does not seem to be supported by java.net.URI.

Other main change has been done in SwaggerLoadSamplesTest that now contains methods to collect model files from either an archive files or local folder. This allow this test to run during maven install phase as well as within Eclipse.