[ZEN-4711] OpenApiDoc: Rendering Error on top-level OAuth Security Scheme
Revert to the change that added the ROOT ObjectType. This change needed to be reverted since it breaks the normailizer by adding unnecessary reprezen extensions to all object nodes.
To solve the problem with security scheme we use a temporary fix that tells the content manager to retain all security schemes. A better solution that trims unused security schemes should be proposed later.