[ZEN-4720] Normalizer: Retain PATH_OR_COMPONENT throws exception
The Option PATH_OR_COMPONENTS could not be parsed by the NormalizerParameters because it was not defined. This is done by adding it in the specialRetainTypes map. Note the document mentions that the option is PATH_OR_COMPONENT in singular but the option in GenFlow is called PATH_OR_COMPONENTS (with components plural). The documentation has been updated to reflect that but I don't know if we should keep the previous one (singular) for backard compatibility.