[ZEN-4489] .gen file missing some SCG and OAG config parameters
This adds support for Swagger/OpenAPI code generators options that are not of type String or Boolean.
These options have be set explicitly to the CodegenConfig object as they cannot be paased to the code generator as additional properties as it is done for
other .gen options that are put in a Map<String, String> and passed to the ClientOptInput object.