[ZEN-4770] Streamline codegen with more granular GenFlow dependencies, fixed-version transitive dependencies
Mark dependencies to eclipse runtime as optional in the genflow-common project.
This project was depending on eclipse runtime plugins for a very limited usage. The class `GeneratedArtifact` was using those to use eclipse IFile. This class  `GeneratedArtifact` and other classes in its package have been removed here because no usage seems to be done in GenFlow or in API Studio. Other usage of eclipse dependencies come from `GenModulesInfo` that uses eclipse FileLocator class to resolve URLs for moduleInfos csv files when GenFlow is used inside eclipse. This dependency has been marked as optional thus avoiding eclipse dependencies to be pulled when GenFlow is used in a non eclipse project.